Iniciando Services

+++ src/main/java/com/senac/votesys/repository/OpcaoVotoRepository.java
@@ -3,5 +3,5 @@ package com.senac.votesys.repository;
 import com.senac.votesys.model.OpcaoVoto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OpcaoVotoRepository extends JpaRepository<OpcaoVoto, Integer> {
+public interface OpcaoVotoRepository extends JpaRepository<OpcaoVoto, Long> {
 }
